+++ HISTORY.md
@@ -25,7 +25,7 @@
     * new "get graphic effect" reporter
     * new "get pen attribute" reporter
     * new "write" command in pen category (used to be "label" in tools)
-    * new "is empty", "map","keep", "combine" and "for each" primitives in list category
+    * new "numbers", "is empty", "map","keep", "combine" and "for each" primitives in list category
     * new "for" loop and "if then else" reporter primitives in the Control category
     * added "neg", "lg" (log2) and "2^" selectors to monadic function reporter in Operators
     * added "^" reporter (power of) in the Operators category
@@ -73,6 +73,7 @@
 
 ### 2019-04-26
 * Lists, Threads, Objects: new "is empty" predicate primitive in List category
+* Threads, Objects: new "numbers" constructor primitive in List category
 
 ### 2019-04-26
 * updated Catalan translation (for new HOF prims)